rename: 패키지명 수정, spot이라는 상위 패키지 생성 후 패키지들 이동
